split out into 2 test files
